addressed Taylor/Rodrigo/Darryls comments
