+++ packages/components/src/Chip/Chip.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { useChildComponent } from "@jobber/hooks/useChildComponent";
 import classNames from "classnames";
 import styles from "./Chip.css";
 import { ChipPrefix } from "./components/ChipPrefix/Chip.Prefix";
 import { ChipSuffix } from "./components/ChipSuffix/Chip.Suffix";
 import { ChipProps } from "./Chip.types";
+import { useChildComponent } from "../hooks/useChildComponent";
 import { Typography } from "../Typography";
 
 export const Chip = ({
@@ -39,13 +39,12 @@ export const Chip = ({
       aria-label={ariaLabel || label}
       disabled={disabled}
       role={role}
+      type="button"
     >
       {prefix}
-      <span>
-        <Typography size="base" fontWeight="medium">
-          {heading}
-        </Typography>
-      </span>
+      <Typography size="base" fontWeight="medium">
+        {heading}
+      </Typography>
       {heading && <span className={styles.chipBar}></span>}
       <span>
         <Typography size="base">{label}</Typography>
