+++ packages/components/src/Chip/Chip.types.d.ts
@@ -1,28 +1,59 @@
 import React, { PropsWithChildren } from "react";
 
 export interface ChipProps extends PropsWithChildren {
-  /** Accessible label, which can be different from the primary label. */
+  /**
+   * Accessible label, which can be different from the primary label.
+   */
   readonly ariaLabel?: string;
-  /** Disables both mouse and keyboard functionality, and updates the visual style of the Chip to appear disabled. */
+
+  /**
+   * Disables both mouse and keyboard functionality, and updates the visual style of the Chip to appear disabled.
+   */
   readonly disabled?: boolean;
-  /** Adds more prominent text to act as a heading. Will be displayed on the left with a | separator. */
+
+  /**
+   * Adds more prominent text to act as a heading. Will be displayed on the left with a | separator.
+   */
   readonly heading?: string;
-  /** Changes Chip styling to inform the user of an issue. */
+
+  /**
+   * Changes Chip styling to inform the user of an issue.
+   */
   readonly invalid?: boolean;
-  /** The content of the chip. Will be displayed on the right if you include a heading. */
+
+  /**
+   * The content of the chip. Will be displayed on the right if you include a heading.
+   */
   readonly label: string;
-  /** The accessible role the Chip is fulfilling. Defaults to 'button' */
+
+  /**
+   * The accessible role the Chip is fulfilling. Defaults to 'button'
+   */
   readonly role?: string;
-  /** Used for accessibility purpopses, specifically using the tab key as navigation. Defaults to 0. */
+
+  /**
+   * Used for accessibility purpopses, specifically using the tab key as navigation.
+   * @default 0
+   */
   readonly tabIndex?: number;
-  /** Button style variation. Does not affect functionality. Defaults to base. */
+
+  /**
+   * Button style variation. Does not affect functionality.
+   * @default "base"
+   */
   readonly variation?: ChipVariations;
-  /** Chip Click Callback. Sends an event and sometimes a value (SelectableChip).  */
+
+  /**
+   * Chip Click Callback. Sends an event and sometimes a value (SelectableChip).
+   */
   readonly onClick?: (
     ev: React.MouseEvent<HTMLButtonElement>,
     value?: string | number,
   ) => void;
-  /**  Callback. Called when you keydown on Chip. Ships the event, so you can get the key pushed.  */
+
+  /**
+   * Callback. Called when you keydown on Chip. Ships the event, so you can get the key pushed.
+   */
   readonly onKeyDown?: (ev: React.KeyboardEvent<HTMLButtonElement>) => void;
 }
