+++ packages/hooks/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./useAssert";
 export * from "./useBreakpoints";
-export * from "./useChildComponent";
 export * from "./useCollectionQuery";
 export * from "./useFocusTrap";
 export * from "./useFormState";
