--- packages/hooks/src/useChildComponent/index.ts
@@ -1,1 +1,0 @@
-export { useChildComponent } from "./useChildComponent";
